feat: allow creating a "flat" structure module
Adds the option "--flat" to the mezzio:module:create command.
When present, the command will not create the nested "src" directory under the module root directory, nor will it create a "templates" directory, either.

Functionality is incomplete; currently, the command proxies to laminas-composer-autoloading, which expects the nested directory when creating autoloading rules.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>