feat: inline ability to register/deregister modules
The laminas-composer-autoloading package only allows the Mezzio "recommended" module structure (i.e., the version that has "src" and "templates" subdirectories).
To allow flat module structures, the mezzio:module:register command needs to be able to detect the type of module present based on the directory layout, and then register it correctly with the package.
Since this functionality is specific to Mezzio, I decided to inline the logic from laminas-composer-autoloading for parsing and updating the `composer.json` file, and to switch to symfony/process for executing the actual `composer` binary for running the `dump-autoload` process.

To make the above testable, I created a few interfaces, as well as default implementations.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>