feat: allow nested namespaces with mezzio:module:create command
The `mezzio:module:create` command now takes an additional option, `--with-namespace` (or `-s`; `-n` is already taken as the short form of `--no-interaction`).
When provided, it prepends that value to the module name being created, separating the two with a namespace separator.
This allows creating new modules with nested namespaces in your application, such as `YourCompany\ModuleName`.

Resolves #1

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>