fix: trim project root from paths pushed into ModuleMetadata
Depending on the system and/or PHP version, `getcwd()` sometimes returns a fully qualified path.
This can cause issues when creating autoloading rules, as paths should be relative to the application root.
The root path is now stripped from the beginning of the project root and module root paths if it is present.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>