Add documentation for the new mezzio:routes:list command to the README
This change adds the command and it's core description to the README,
and also adds an extended documentation section to the end of the
README, showing what the command does, the options it supports, and some
sample output. The tabular output isn't quite correct, but is there to
at least have a placeholder.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>