Add the core classes for the command
This change adds the core classes, ListRoutesCommand and
ListRoutesCommandFactory, along with accompanying tests for them. At
this stage, the functionality is extremely limited, covering that the
command can be instantiated correctly, and has the correct
arguments/options, help output strings, command name, and description.
The rest of the functionality is coming.