Replace the RouteCollector member variable with an array of Routes
I don't know why I didn't do this originally, but there was no real need
to have a RouteCollector variable (that I know of) as a member variable.
All the application needs is an array of Route objects, as that's all it
iterates over. This change corrects the original oversight.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>