Finish implementing the routing filter options
This change finishes implementing the routing filter options, so that
the options passed from the command line when the command is called,
that those options will be respected. It also does a little bit of code
clean up so that the code is easier to read and maintain (I hope).

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>