Add bin property to composer.json
Without this, when running vendor/bin/mezzio-tooling in projects that
use this package, the mezzio command namespace is missing when running
vendor/bin/laminas. As reported in #30, the following error is printed:

Installed tooling package, but package does not contain script information?
Check your installation, and follow the migration documentation to use the tooling.

After some experimentation, by adding this addition to the
configuration, the error isn't displayed and the namespace is once again
available.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>