Add the ability to list an application's routing table
This was originally created in #27, but that PR got quite messy. So, my
intention is to add the code in a new PR with the baggage of the
original.

The change, in short, adds a new Mezzio command: routes:list. When
invoked, it will print the application's routing table akin to what
Symfony and Laravel provide.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>