Refactor RoutesFilter to use a RouteFilterOptionsInterface object
Previously, RoutesFilter was taking an array with the filter options
(i.e., method, path, name, etc) but, as @gsteel noted, there was a
perfectly good RouteFilterOptions class available. I don't know why I
developed it, but overlooked it. This change refactors RoutesFilter to
use it instead, then goes one step further and introduces a
RouteFilterOptionsInterface so that, when no filter options are
available, an EmptyRouteFilterOptions can be supplied, hopefully making
the code easier to intuit and maintain.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>