Refactor how ListRoutesCommandFactory retrieves a ConfigLoaderInterface
This change refactors ListRoutesCommandFactory to retrieve a
ConfigLoaderInterface object from the application's DiC instead of
instantiating a RoutesFileConfigLoader object directly. The main reason
for this is to provide flexibility to the user, should they have the
need to retrieve routes in a different way than the default
implementation, which searches in config/routes.php and in any routes
registered through ConfigProvider classes.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>