Add a default way of retrieving routes from the application
This change provides a default implementation for retrieving an
application's routes configuration. It uses the stock approach of
searching for routes in config/routes.php and any registered with the
Application object in any of the loaded ConfigProvider classes. It also
provides a simplistic example for building a custom implementation,
should it be required.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>